Increased flash z-index to 2147483647
I've recently had problems getting FileAPI to use Flash. Although Flash was being detected and the flash uploader inserted on the page, FileAPI was falling back to using an iFrame for submission. After investigating further, it turns out that the flash uploader was being hidden behind other elements. FileAPI uses a z-index of 1000000, but apparently some of our elements are even higher than that (don't ask why). To resolve this, I've increased the z-index to 2^32, which I believe is the maximum that both legacy and new browsers will accept (see http://www.puidokas.com/max-z-index/). This resolves the problem for me, and now FileAPI successfully uses the flash uploader.